requirements.txt
moviepy
openai-whisper
pandas
pydub
scipy